+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
+### Fixed
+- `pw.xpacks.llm.document_store.DocumentStore` no longer requires `_metadata` column in the input table.
+
 ## [0.16.0] - 2024-11-29
 
 ### Added